requirements.txt
Django>=5.0
django-unfold>=0.16.0
django-mdeditor>=0.1.20
Pillow>=10.0.0
markdown>=3.4
gunicorn>=21.2.0
python-dotenv>=1.0.0
whitenoise>=6.6.0
